+++ repology_client/__init__.py
@@ -1,13 +1,16 @@
 # SPDX-License-Identifier: EUPL-1.2
-# SPDX-FileCopyrightText: 2024 Anna <cyber@sysrq.in>
+# SPDX-FileCopyrightText: 2024-2025 Anna <cyber@sysrq.in>
 
-""" Asynchronous wrapper for Repology API. """
+"""
+Asynchronous wrapper for Repology API.
+"""
 
 from types import SimpleNamespace
 
 from repology_client._client.v1 import (
     api,
     get_packages,
+    get_problems,
     get_projects,
 )
 from repology_client._client.tools import (
@@ -27,6 +30,7 @@ exp.distromap = distromap
 __all__ = [
     "api",
     "get_packages",
+    "get_problems",
     "get_projects",
     "resolve_package",